requirements.txt
aiohttp==3.8.1
asgiref==3.5.0
fastapi==0.76.0
geopandas==0.10.2
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
pydantic==1.9.0
pyproj==3.3.0
python-dotenv==0.20.0
pytorch_lightning==1.6.1
scikit_learn==1.0.2
seaborn==0.11.2
Shapely==1.8.1.post1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
uvicorn==0.17.6
hydra-core